# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HST API Documentation",
        "description": "Application API services documentation page",
        "contact": {
            "email": "volkangurgor@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/teachers": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Get all teachers",
                "operationId": "24a0f742e0dce09dcf48f367d3b24295",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Create a new teacher",
                "operationId": "814d9c14a01fe7875bc7c8fe2a1fa70e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "majors": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/teachers/{id}": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Get a specific teacher",
                "operationId": "536660906477ddbb95d1854c9eceeb51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Teacher ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Teacher not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Update a teacher",
                "operationId": "2e65d9a5d98ae4d6f47ee82d709d12f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Teacher ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "majors": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Teacher not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Delete a teacher",
                "operationId": "34c7b4366727e061a05878d3895d3dd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Teacher ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Teacher not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "name": "Authorization",
                "description": "Bearer token to access these API endpoints",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}